Create Hashing
This Java program demonstrates the use of the HashMap class from the java.util package to store and manipulate key-value pairs. The program performs the following operations:

HashMap Initialization:

A HashMap is created to store student IDs (Integer) as keys and student names (String) as values.
Adding Entries:

The put() method is used to add key-value pairs to the HashMap.
Retrieving Values:

The values() method is used to retrieve a Collection of all values (student names) in the HashMap.
Iterating Over Entries:

The entrySet() method is used to retrieve a Set of all key-value pairs in the HashMap.
A for-each loop is used to iterate over the entries and print the key-value pairs.

Removing Entries:

Keys to be removed are collected in a List during iteration and removed from the HashMap after the iteration.
Replacing and Adding Entries:

The replace() method is used to update the value of an existing key.
The putIfAbsent() method is used to add a new key-value pair only if the key does not already exist.
Final State:

The program prints the final state of the HashMap, its size, and whether it is empty.